requirements.txt
# Core Django Framework
Django==5.2.3
django-allauth==65.9.0

# RESTful API Framework (for external integrations)
djangorestframework==3.15.2

# Security and Cryptography
cryptography==46.0.3
cffi==2.0.0
PyJWT==2.10.1

# UI and Form Enhancement
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
django-colorfield==0.14.0
django-widget-tweaks==1.5.0

# Image Processing
Pillow==12.0.0

# Environment and Configuration
python-dotenv==1.1.0

# Production Server
gunicorn==23.0.0
whitenoise==6.9.0

# AI Integration - Google Gemini
google-generativeai==0.8.5
grpcio==1.76.0
pydantic==2.12.3
pydantic-core==2.41.4

# Document Processing
python-docx==1.2.0
PyPDF2==3.0.1
Markdown==3.5.1

# HTTP Requests (for testing and integrations)
requests==2.32.3

# Real-time Communication (WebSockets)
channels==4.1.0
channels-redis==4.1.0
daphne==4.1.2

# Asynchronous Task Processing
celery==5.3.4
redis==5.0.1

# Development and Testing Dependencies
# (Optional - for development environment)
# flake8==7.1.1
# black==24.10.0
# isort==5.13.2